+++ pkg/scalers/pulsar_scaler_test.go
@@ -20,25 +20,21 @@ type parsePulsarMetadataTestData struct {
 }
 
 type parsePulsarAuthParamsTestData struct {
-	authParams map[string]string
-	isError    bool
-	enableTLS  bool
-	cert       string
-	key        string
-	ca         string
+	triggerMetadata map[string]string
+	authParams      map[string]string
+	isError         bool
+	enableTLS       bool
+	cert            string
+	key             string
+	ca              string
+	bearerToken     string
+	username        string
+	password        string
 }
 
 type pulsarMetricIdentifier struct {
 	metadataTestData *parsePulsarMetadataTestData
 	name             string
-}
-
-// A complete valid metadata example for reference
-var validPulsarMetadata = map[string]string{
-	"adminURL":     "http://172.20.0.151:80",
-	"topic":        "persistent://public/default/my-topic",
-	"subscription": "sub1",
-	"tls":          "enable",
 }
 
 // A complete valid authParams example for sasl, with username and passwd
@@ -70,8 +66,19 @@ var parsePulsarMetadataTestDataset = []parsePulsarMetadataTestData{
 }
 
 var parsePulsarMetadataTestAuthTLSDataset = []parsePulsarAuthParamsTestData{
-	// failure, no adminURL
-	{map[string]string{"cert": "certdata", "key": "keydata", "ca": "cadata"}, false, true, "certdata", "keydata", "cadata"},
+	// Passes, mutual TLS, no other auth (legacy "tls: enable")
+	{map[string]string{"adminURL": "http://172.20.0.151:80", "topic": "persistent://public/default/my-topic", "subscription": "sub1", "tls": "enable"}, map[string]string{"cert": "certdata", "key": "keydata", "ca": "cadata"}, false, true, "certdata", "keydata", "cadata", "", "", ""},
+	// Passes, mutual TLS, no other auth (uses new way to enable tls)
+	{map[string]string{"adminURL": "http://172.20.0.151:80", "topic": "persistent://public/default/my-topic", "subscription": "sub1", "authModes": "tls"}, map[string]string{"cert": "certdata", "key": "keydata", "ca": "cadata"}, false, true, "certdata", "keydata", "cadata", "", "", ""},
+	// Fails, mutual TLS (legacy "tls: enable") without cert
+	{map[string]string{"adminURL": "http://172.20.0.151:80", "topic": "persistent://public/default/my-topic", "subscription": "sub1", "tls": "enable"}, map[string]string{"cert": "", "key": "keydata", "ca": "cadata"}, true, true, "certdata", "keydata", "cadata", "", "", ""},
+	// Fails, mutual TLS, (uses new way to enable tls) without cert
+	{map[string]string{"adminURL": "http://172.20.0.151:80", "topic": "persistent://public/default/my-topic", "subscription": "sub1", "authModes": "tls"}, map[string]string{"cert": "certdata", "key": "", "ca": "cadata"}, true, true, "certdata", "keydata", "cadata", "", "", ""},
+	// Passes, server side TLS with bearer token. Note that EnableTLS is expected to be false because it is not mTLS.
+	// The legacy behavior required tls: enable in order to configure a custom root ca. Now, all that is required is configuring a root ca.
+	{map[string]string{"adminURL": "http://172.20.0.151:80", "topic": "persistent://public/default/my-topic", "subscription": "sub1", "tls": "enable", "authModes": "bearer"}, map[string]string{"ca": "cadata", "bearerToken": "my-special-token"}, false, false, "", "", "cadata", "my-special-token", "", ""},
+	// Passes, server side TLS with basic auth. Note that EnableTLS is expected to be false because it is not mTLS.
+	{map[string]string{"adminURL": "http://172.20.0.151:80", "topic": "persistent://public/default/my-topic", "subscription": "sub1", "authModes": "basic"}, map[string]string{"ca": "cadata", "username": "admin", "password": "password123"}, false, false, "", "", "cadata", "", "admin", "password123"},
 }
 
 var pulsarMetricIdentifiers = []pulsarMetricIdentifier{
@@ -143,28 +150,54 @@ func TestParsePulsarMetadata(t *testing.T) {
 
 func TestPulsarAuthParams(t *testing.T) {
 	for _, testData := range parsePulsarMetadataTestAuthTLSDataset {
-		meta, err := parsePulsarMetadata(&ScalerConfig{TriggerMetadata: validPulsarMetadata, AuthParams: testData.authParams})
+		meta, err := parsePulsarMetadata(&ScalerConfig{TriggerMetadata: testData.triggerMetadata, AuthParams: testData.authParams})
 
 		if err != nil && !testData.isError {
-			t.Error("Expected success but got error", err)
+			t.Error("Expected success but got error", testData.authParams, err)
 		}
 		if testData.isError && err == nil {
 			t.Error("Expected error but got success")
 		}
-		if meta.enableTLS != testData.enableTLS {
-			t.Errorf("Expected enableTLS to be set to %v but got %v\n", testData.enableTLS, meta.enableTLS)
+
+		if meta.pulsarAuth == nil {
+			t.Log("meta.pulsarAuth is nil, skipping rest of validation of", testData)
+			continue
 		}
 
-		if meta.ca != testData.ca {
-			t.Errorf("Expected ca to be set to %s but got %s\n", testData.ca, meta.ca)
+		if meta.pulsarAuth.EnableTLS != testData.enableTLS {
+			t.Errorf("Expected enableTLS to be set to %v but got %v\n", testData.enableTLS, meta.pulsarAuth.EnableTLS)
 		}
 
-		if meta.cert != testData.cert {
-			t.Errorf("Expected cert to be set to %s but got %s\n", testData.cert, meta.cert)
+		if meta.pulsarAuth.CA != testData.ca {
+			t.Errorf("Expected ca to be set to %s but got %s\n", testData.ca, meta.pulsarAuth.CA)
 		}
 
-		if meta.key != testData.key {
-			t.Errorf("Expected key to be set to %s but got %s\n", testData.key, meta.key)
+		if meta.pulsarAuth.Cert != testData.cert {
+			t.Errorf("Expected cert to be set to %s but got %s\n", testData.cert, meta.pulsarAuth.Cert)
+		}
+
+		if meta.pulsarAuth.Key != testData.key {
+			t.Errorf("Expected key to be set to %s but got %s\n", testData.key, meta.pulsarAuth.Key)
+		}
+
+		if meta.pulsarAuth.EnableBearerAuth != (testData.bearerToken != "") {
+			t.Errorf("Expected EnableBearerAuth to be true when bearerToken is %s\n", testData.bearerToken)
+		}
+
+		if meta.pulsarAuth.BearerToken != testData.bearerToken {
+			t.Errorf("Expected bearer token to be set to %s but got %s\n", testData.bearerToken, meta.pulsarAuth.BearerToken)
+		}
+
+		if meta.pulsarAuth.EnableBasicAuth != (testData.username != "" || testData.password != "") {
+			t.Errorf("Expected EnableBearerAuth to be true when bearerToken is %s\n", testData.bearerToken)
+		}
+
+		if meta.pulsarAuth.Username != testData.username {
+			t.Errorf("Expected username to be set to %s but got %s\n", testData.username, meta.pulsarAuth.Username)
+		}
+
+		if meta.pulsarAuth.Password != testData.password {
+			t.Errorf("Expected password to be set to %s but got %s\n", testData.password, meta.pulsarAuth.Password)
 		}
 	}
 }